creates TeamOutputPage
